k/server: Split out topic config utils
