+++ src/v/kafka/server/BUILD
@@ -35,6 +35,51 @@ redpanda_cc_library(
 )
 
 redpanda_cc_library(
+    name = "topic_config_utils",
+    srcs = [
+        "handlers/configs/config_response_utils.cc",
+        "handlers/topics/topic_utils.cc",
+        "handlers/topics/types.cc",
+    ],
+    hdrs = [
+        "handlers/configs/config_response_utils.h",
+        "handlers/configs/config_utils.h",
+        "handlers/topics/topic_utils.h",
+        "handlers/topics/types.h",
+        "handlers/topics/validators.h",
+    ],
+    include_prefix = "kafka/server",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":errors",
+        "//src/v/base",
+        "//src/v/cluster",
+        "//src/v/config",
+        "//src/v/container:chunked_vector",
+        "//src/v/datalake:partition_spec_parser",
+        "//src/v/kafka/protocol",
+        "//src/v/kafka/protocol:describe_configs",
+        "//src/v/kafka/protocol:logger",
+        "//src/v/kafka/protocol/schemata:create_topics_request",
+        "//src/v/kafka/protocol/schemata:create_topics_response",
+        "//src/v/model",
+        "//src/v/pandaproxy/schema_registry:config",
+        "//src/v/pandaproxy/schema_registry:subject_name_strategy",
+        "//src/v/security",
+        "//src/v/serde:chrono",
+        "//src/v/strings:string_switch",
+        "//src/v/utils:absl_sstring_hash",
+        "//src/v/utils:tristate",
+        "@abseil-cpp//absl/container:flat_hash_map",
+        "@abseil-cpp//absl/container:node_hash_set",
+        "@boost//:algorithm",
+        "@boost//:container",
+        "@boost//:lexical_cast",
+        "@seastar",
+    ],
+)
+
+redpanda_cc_library(
     name = "server",
     srcs = [
         "client_quota_translator.cc",
@@ -57,7 +102,6 @@ redpanda_cc_library(
         "handlers/alter_user_scram_credentials.cc",
         "handlers/api_versions.cc",
         "handlers/client_quotas.cc",
-        "handlers/configs/config_response_utils.cc",
         "handlers/create_partitions.cc",
         "handlers/create_topics.cc",
         "handlers/delete_acls.cc",
@@ -79,8 +123,6 @@ redpanda_cc_library(
         "handlers/metadata.cc",
         "handlers/offset_for_leader_epoch.cc",
         "handlers/produce.cc",
-        "handlers/topics/topic_utils.cc",
-        "handlers/topics/types.cc",
         "handlers/txn_offset_commit.cc",
         "member.cc",
         "protocol_utils.cc",
@@ -124,8 +166,6 @@ redpanda_cc_library(
         "handlers/alter_partition_reassignments.h",
         "handlers/alter_user_scram_credentials.h",
         "handlers/api_versions.h",
-        "handlers/configs/config_response_utils.h",
-        "handlers/configs/config_utils.h",
         "handlers/create_acls.h",
         "handlers/create_partitions.h",
         "handlers/create_topics.h",
@@ -142,6 +182,7 @@ redpanda_cc_library(
         "handlers/describe_producers.h",
         "handlers/describe_transactions.h",
         "handlers/describe_user_scram_credentials.h",
+        "handlers/details/alter_config_utils.h",
         "handlers/details/leader_epoch.h",
         "handlers/details/security.h",
         "handlers/end_txn.h",
@@ -172,9 +213,6 @@ redpanda_cc_library(
         "handlers/sasl_authenticate.h",
         "handlers/sasl_handshake.h",
         "handlers/sync_group.h",
-        "handlers/topics/topic_utils.h",
-        "handlers/topics/types.h",
-        "handlers/topics/validators.h",
         "handlers/txn_offset_commit.h",
         "kafka_probe.h",
         "member.h",
@@ -202,6 +240,7 @@ redpanda_cc_library(
         ":datalake_usage_api",
         ":logger",
         ":qdc_monitor_config",
+        ":topic_config_utils",
         "//src/v/base",
         "//src/v/bytes",
         "//src/v/bytes:iobuf",
