+++ src/v/kafka/server/tests/BUILD
@@ -122,6 +122,7 @@ redpanda_cc_btest(
         "//src/v/cluster",
         "//src/v/kafka/protocol/schemata:create_topics_request",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/model",
         "//src/v/test_utils:seastar_boost",
         "@boost//:test",
@@ -138,6 +139,7 @@ redpanda_cc_btest(
     deps = [
         "//src/v/kafka/protocol/schemata:create_topics_response",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/model",
         "//src/v/test_utils:seastar_boost",
         "@boost//:range",
@@ -169,6 +171,7 @@ redpanda_cc_btest(
     deps = [
         "//src/v/kafka/protocol/schemata:create_topics_request",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/pandaproxy/schema_registry:subject_name_strategy",
         "//src/v/test_utils:seastar_boost",
         "@boost//:test",
@@ -202,6 +205,7 @@ redpanda_cc_btest(
         "//src/v/cluster",
         "//src/v/config",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/model",
         "//src/v/test_utils:seastar_boost",
         "@boost//:test",
@@ -349,6 +353,7 @@ redpanda_cc_btest(
         "//src/v/kafka/protocol:metadata",
         "//src/v/kafka/protocol/schemata:create_topics_request",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/redpanda:application",
         "//src/v/redpanda/tests:fixture",
         "//src/v/test_utils:seastar_boost",
@@ -422,6 +427,7 @@ redpanda_cc_btest(
         "//src/v/kafka/protocol:create_topics",
         "//src/v/kafka/protocol:metadata",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/kafka/server/tests:topic_properties_helpers",
         "//src/v/model",
         "//src/v/test_utils:seastar_boost",
@@ -482,6 +488,7 @@ redpanda_cc_btest(
         "//src/v/kafka/protocol:metadata",
         "//src/v/kafka/protocol:produce",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/model",
         "//src/v/model/tests:random",
         "//src/v/redpanda:application",
@@ -572,6 +579,7 @@ redpanda_cc_btest(
         "//src/v/kafka/protocol/schemata:alter_partition_reassignments_request",
         "//src/v/kafka/protocol/schemata:list_partition_reassignments_request",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/model",
         "//src/v/redpanda/tests:fixture",
         "//src/v/test_utils:seastar_boost",
@@ -682,6 +690,7 @@ redpanda_cc_btest(
         "//src/v/kafka/protocol:incremental_alter_configs",
         "//src/v/kafka/protocol:metadata",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:topic_config_utils",
         "//src/v/kafka/server/tests:topic_properties_helpers",
         "//src/v/model",
         "//src/v/test_utils:scoped_config",
